pyproject.toml
[project]
name = "rag-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.1",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "load-dotenv>=0.1.0",
    "openai>=2.16.0",
    "qdrant-client>=1.16.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
